WIP LoginCommand.

+++ src/Domain/Entities/OperatorTerminal.cs
@@ -3,7 +3,7 @@ public class OperatorTerminal
 {
     public int Id { get; set; }
 
-    public int OperatorId { get; set; }
+    public short OperatorId { get; set; }
     public virtual Operator Operator { get; set; } = null!;
 
     public int TerminalId { get; set; }
